refactored_marketing_agent
