+++ REFACTORING.md
@@ -129,6 +129,7 @@ When refactoring an agent:
 | `blog_agent_service.py` | 534 lines | 202 lines | `blog_tool_executor.py` (232 lines) | Done |
 | `website_agent_service.py` | 760 lines | 197 lines | `website_tool_executor.py` (338 lines) | Done |
 | `business_report_agent_service.py` | 619 lines | 299 lines | `business_report_tool_executor.py` (280 lines) | Done |
+| `marketing_strategy_agent_service.py` | 489 lines | 202 lines | `marketing_strategy_tool_executor.py` (230 lines) | Done |
 
 ---
 
@@ -144,7 +145,7 @@ When refactoring an agent:
 
 | Utility | Location | Purpose | Used By |
 |---------|----------|---------|---------|
-| `get_source_content()` | `app/utils/source_content_utils.py` | Load source content with smart sampling for large sources | blog_agent, website_agent, wireframe_service |
+| `get_source_content()` | `app/utils/source_content_utils.py` | Load source content with smart sampling for large sources | blog_agent, website_agent, wireframe_service, marketing_strategy_agent |
 | `get_source_name()` | `app/utils/source_content_utils.py` | Get source name by ID | (available) |
 | `convert_to_excalidraw_elements()` | `app/utils/excalidraw_utils.py` | Convert simplified elements to Excalidraw format | wireframe_service |
