+++ backend/app/services/ai_agents/marketing_strategy_agent_service.py
@@ -1,15 +1,12 @@
 """
 Marketing Strategy Agent Service - AI agent for generating Marketing Strategy Documents.
 
-Educational Note: Agentic loop pattern for document generation:
+Orchestrates the marketing strategy generation workflow:
 1. Agent plans the strategy structure (plan_marketing_strategy tool)
-2. Agent writes sections incrementally (write_marketing_section tool - can be called multiple times)
+2. Agent writes sections incrementally (write_marketing_section tool)
 3. Agent signals completion via is_last_section=true flag
-
-The markdown output can be rendered on frontend and exported to PDF.
 """
 
-import os
 import uuid
 from typing import Dict, Any, List
 from datetime import datetime
@@ -17,42 +14,31 @@ from datetime import datetime
 from app.services.integrations.claude import claude_service
 from app.config import prompt_loader, tool_loader
 from app.utils import claude_parsing_utils
-from app.utils.path_utils import get_studio_dir, get_sources_dir
+from app.utils.source_content_utils import get_source_content
 from app.services.data_services import message_service
 from app.services.studio_services import studio_index_service
+from app.services.tool_executors.marketing_strategy_tool_executor import marketing_strategy_tool_executor
 
 
 class MarketingStrategyAgentService:
-    """
-    Marketing strategy generation agent with multi-step document writing workflow.
-
-    Educational Note: This agent demonstrates how AI can create structured
-    documents incrementally: planning -> writing sections -> completion.
-    """
+    """Marketing strategy generation agent - orchestration only."""
 
     AGENT_NAME = "marketing_strategy_agent"
-    MAX_ITERATIONS = 10  # Brief strategies: 1 plan + ~5 sections
+    MAX_ITERATIONS = 10
 
     def __init__(self):
-        """Initialize agent with lazy-loaded config and tools."""
         self._prompt_config = None
         self._tools = None
 
     def _load_config(self) -> Dict[str, Any]:
-        """Lazy load prompt configuration."""
         if self._prompt_config is None:
             self._prompt_config = prompt_loader.get_prompt_config("marketing_strategy_agent")
         return self._prompt_config
 
     def _load_tools(self) -> List[Dict[str, Any]]:
-        """Load all agent tools."""
         if self._tools is None:
             self._tools = tool_loader.load_tools_for_agent(self.AGENT_NAME)
         return self._tools
-
-    # =========================================================================
-    # Main Agent Execution
-    # =========================================================================
 
     def generate_marketing_strategy(
         self,
@@ -61,16 +47,7 @@ class MarketingStrategyAgentService:
         job_id: str,
         direction: str = ""
     ) -> Dict[str, Any]:
-        """
-        Run the agent to generate a marketing strategy document.
-
-        Educational Note: The agent workflow:
-        1. Get source content and direction
-        2. Agent plans the document (sections, structure)
-        3. Agent writes sections incrementally to markdown file
-        4. Agent signals completion with is_last_section=true
-        5. We finalize and update job status
-        """
+        """Run the agent to generate a marketing strategy document."""
         config = self._load_config()
         tools = self._load_tools()
 
@@ -85,36 +62,27 @@ class MarketingStrategyAgentService:
             started_at=started_at
         )
 
-        # Get source content
-        source_content = self._get_source_content(project_id, source_id)
+        # Get source content using shared utility
+        source_content = get_source_content(project_id, source_id, max_chars=15000)
 
-        # Build initial user message
-        user_message = f"""Create a comprehensive Marketing Strategy Document based on the following source content.
-
-=== SOURCE CONTENT ===
-{source_content}
-=== END SOURCE CONTENT ===
-
-Direction from user: {direction if direction else 'No specific direction provided - create a complete marketing strategy covering all relevant aspects of the product/service.'}
-
-Please create a complete marketing strategy following the workflow:
-1. First, plan the document structure using the plan_marketing_strategy tool
-2. Then write each section one at a time using the write_marketing_section tool
-3. Set is_last_section=true when you write the final section"""
+        # Build user message from config
+        effective_direction = direction if direction else config.get("default_direction", "")
+        user_message = config.get("user_message", "").format(
+            source_content=source_content,
+            direction=effective_direction
+        )
 
         messages = [{"role": "user", "content": user_message}]
 
         total_input_tokens = 0
         total_output_tokens = 0
         sections_written = 0
-        markdown_file_path = None
 
         print(f"[MarketingStrategyAgent] Starting (job_id: {job_id[:8]})")
 
         for iteration in range(1, self.MAX_ITERATIONS + 1):
             print(f"  Iteration {iteration}/{self.MAX_ITERATIONS}")
 
-            # Call Claude API
             response = claude_service.send_message(
                 messages=messages,
                 system_prompt=config["system_prompt"],
@@ -126,11 +94,9 @@ Please create a complete marketing strategy following the workflow:
                 project_id=project_id
             )
 
-            # Track token usage
             total_input_tokens += response["usage"]["input_tokens"]
             total_output_tokens += response["usage"]["output_tokens"]
 
-            # Serialize and add assistant response to messages
             content_blocks = response.get("content_blocks", [])
             serialized_content = claude_parsing_utils.serialize_content_blocks(content_blocks)
             messages.append({"role": "assistant", "content": serialized_content})
@@ -148,48 +114,41 @@ Please create a complete marketing strategy following the workflow:
 
                     print(f"    Tool: {tool_name}")
 
-                    # Tool 1: Plan the marketing strategy
-                    if tool_name == "plan_marketing_strategy":
-                        result = self._handle_plan_marketing_strategy(project_id, job_id, tool_input)
-                        tool_results.append({
-                            "type": "tool_result",
-                            "tool_use_id": tool_id,
-                            "content": result
-                        })
+                    # Build execution context
+                    context = {
+                        "project_id": project_id,
+                        "job_id": job_id,
+                        "source_id": source_id,
+                        "sections_written": sections_written,
+                        "iterations": iteration,
+                        "input_tokens": total_input_tokens,
+                        "output_tokens": total_output_tokens
+                    }
 
-                    # Tool 2: Write marketing section (also handles termination)
-                    elif tool_name == "write_marketing_section":
-                        result, is_complete, file_path = self._handle_write_section(
-                            project_id, job_id, tool_input, sections_written
-                        )
+                    # Execute tool via executor
+                    result, is_termination = marketing_strategy_tool_executor.execute_tool(
+                        tool_name, tool_input, context
+                    )
+
+                    # Track sections written
+                    if tool_name == "write_marketing_section":
                         sections_written += 1
-                        if file_path:
-                            markdown_file_path = file_path
 
-                        tool_results.append({
-                            "type": "tool_result",
-                            "tool_use_id": tool_id,
-                            "content": result
-                        })
+                    if is_termination:
+                        print(f"  Completed in {iteration} iterations, {sections_written} sections")
+                        self._save_execution(
+                            project_id, execution_id, job_id, messages,
+                            result, started_at, source_id
+                        )
+                        return result
 
-                        # Check if this was the last section (TERMINATION)
-                        if is_complete:
-                            final_result = self._finalize_marketing_strategy(
-                                project_id, job_id, source_id, markdown_file_path,
-                                sections_written, iteration, total_input_tokens, total_output_tokens
-                            )
+                    # Add tool result
+                    tool_results.append({
+                        "type": "tool_result",
+                        "tool_use_id": tool_id,
+                        "content": result.get("message", str(result))
+                    })
 
-                            print(f"  Completed in {iteration} iterations, {sections_written} sections")
-
-                            # Save execution log
-                            self._save_execution(
-                                project_id, execution_id, job_id, messages,
-                                final_result, started_at, source_id
-                            )
-
-                            return final_result
-
-            # Add tool results to messages
             if tool_results:
                 messages.append({"role": "user", "content": tool_results})
 
@@ -216,251 +175,6 @@ Please create a complete marketing strategy following the workflow:
 
         return error_result
 
-    # =========================================================================
-    # Tool Handlers
-    # =========================================================================
-
-    def _handle_plan_marketing_strategy(
-        self,
-        project_id: str,
-        job_id: str,
-        tool_input: Dict[str, Any]
-    ) -> str:
-        """Handle plan_marketing_strategy tool call."""
-        document_title = tool_input.get("document_title", "Marketing Strategy Document")
-        product_name = tool_input.get("product_name", "Unknown Product")
-        sections = tool_input.get("sections", [])
-
-        print(f"      Planning: {document_title} ({len(sections)} sections)")
-
-        # Update job with plan
-        studio_index_service.update_marketing_strategy_job(
-            project_id, job_id,
-            document_title=document_title,
-            product_name=product_name,
-            target_market=tool_input.get("target_market"),
-            planned_sections=sections,
-            planning_notes=tool_input.get("planning_notes"),
-            total_sections=len(sections),
-            status_message=f"Planned {len(sections)} sections, starting to write..."
-        )
-
-        return f"Marketing strategy plan saved successfully. Document: '{document_title}', Product: '{product_name}', Sections planned: {len(sections)}. Now proceed to write each section using the write_marketing_section tool."
-
-    def _handle_write_section(
-        self,
-        project_id: str,
-        job_id: str,
-        tool_input: Dict[str, Any],
-        current_sections_written: int
-    ) -> tuple:
-        """
-        Handle write_marketing_section tool call.
-
-        Educational Note: We use our own counter (current_sections_written + 1)
-        instead of trusting the LLM's section_number. This prevents duplicate
-        sections if the LLM sends the same section_number repeatedly.
-
-        Returns:
-            tuple: (result_message, is_complete, file_path)
-        """
-        # Use our own counter - don't trust LLM's section_number to avoid duplicates
-        actual_section_number = current_sections_written + 1
-        agent_section_number = tool_input.get("section_number", actual_section_number)
-
-        operation = tool_input.get("operation", "append")
-        is_last_section = tool_input.get("is_last_section", False)
-        section_title = tool_input.get("section_title", "")
-        markdown_content = tool_input.get("markdown_content", "")
-
-        # Log if there's a mismatch (for debugging)
-        if agent_section_number != actual_section_number:
-            print(f"      Note: Agent sent section {agent_section_number}, using actual count {actual_section_number}")
-
-        print(f"      Writing section {actual_section_number}: {section_title} (is_last: {is_last_section})")
-
-        try:
-            # Prepare output directory
-            studio_dir = get_studio_dir(project_id)
-            marketing_strategy_dir = os.path.join(studio_dir, "marketing_strategies")
-            os.makedirs(marketing_strategy_dir, exist_ok=True)
-
-            # File path
-            markdown_filename = f"{job_id}.md"
-            file_path = os.path.join(marketing_strategy_dir, markdown_filename)
-
-            # Get job info for document title and total sections
-            job = studio_index_service.get_marketing_strategy_job(project_id, job_id)
-            document_title = job.get("document_title", "Marketing Strategy Document") if job else "Marketing Strategy Document"
-            total_sections = job.get("total_sections", 0) if job else 0
-
-            # Write or append content
-            if operation == "write":
-                # First section - create file with title
-                with open(file_path, "w", encoding="utf-8") as f:
-                    f.write(f"# {document_title}\n\n")
-                    f.write(f"*Generated on {datetime.now().strftime('%Y-%m-%d %H:%M')}*\n\n")
-                    f.write("---\n\n")
-                    f.write(markdown_content)
-                    f.write("\n\n")
-            else:
-                # Subsequent sections - append
-                with open(file_path, "a", encoding="utf-8") as f:
-                    f.write(markdown_content)
-                    f.write("\n\n")
-
-            studio_index_service.update_marketing_strategy_job(
-                project_id, job_id,
-                sections_written=actual_section_number,
-                current_section=section_title,
-                markdown_file=markdown_filename,
-                status_message=f"Writing section {actual_section_number}/{total_sections}: {section_title}..."
-            )
-
-            # Provide clear feedback to help Claude know what to do next
-            result_msg = f"Section {actual_section_number} '{section_title}' written successfully."
-            if is_last_section:
-                result_msg += " Marketing strategy document is now complete."
-            elif total_sections > 0:
-                remaining = total_sections - actual_section_number
-                result_msg += f" Progress: {actual_section_number}/{total_sections} sections complete. {remaining} section(s) remaining."
-
-            return result_msg, is_last_section, file_path
-
-        except Exception as e:
-            error_msg = f"Error writing section {actual_section_number}: {str(e)}"
-            print(f"      {error_msg}")
-            return error_msg, False, None
-
-    def _finalize_marketing_strategy(
-        self,
-        project_id: str,
-        job_id: str,
-        source_id: str,
-        file_path: str,
-        sections_written: int,
-        iterations: int,
-        input_tokens: int,
-        output_tokens: int
-    ) -> Dict[str, Any]:
-        """Finalize the marketing strategy document and update job status."""
-        try:
-            # Get job info
-            job = studio_index_service.get_marketing_strategy_job(project_id, job_id)
-            document_title = job.get("document_title", "Marketing Strategy") if job else "Marketing Strategy"
-            markdown_filename = f"{job_id}.md"
-
-            # Update job to ready
-            studio_index_service.update_marketing_strategy_job(
-                project_id, job_id,
-                status="ready",
-                status_message="Marketing strategy generated successfully!",
-                markdown_file=markdown_filename,
-                markdown_filename=markdown_filename,
-                preview_url=f"/api/v1/projects/{project_id}/studio/marketing-strategies/{job_id}/preview",
-                download_url=f"/api/v1/projects/{project_id}/studio/marketing-strategies/{job_id}/download",
-                iterations=iterations,
-                input_tokens=input_tokens,
-                output_tokens=output_tokens,
-                completed_at=datetime.now().isoformat()
-            )
-
-            return {
-                "success": True,
-                "job_id": job_id,
-                "document_title": document_title,
-                "markdown_file": markdown_filename,
-                "preview_url": f"/api/v1/projects/{project_id}/studio/marketing-strategies/{job_id}/preview",
-                "download_url": f"/api/v1/projects/{project_id}/studio/marketing-strategies/{job_id}/download",
-                "sections_written": sections_written,
-                "iterations": iterations,
-                "usage": {"input_tokens": input_tokens, "output_tokens": output_tokens}
-            }
-
-        except Exception as e:
-            error_msg = f"Error finalizing marketing strategy: {str(e)}"
-            print(f"      {error_msg}")
-
-            studio_index_service.update_marketing_strategy_job(
-                project_id, job_id,
-                status="error",
-                error_message=error_msg
-            )
-
-            return {
-                "success": False,
-                "error_message": error_msg,
-                "iterations": iterations,
-                "usage": {"input_tokens": input_tokens, "output_tokens": output_tokens}
-            }
-
-    # =========================================================================
-    # Helper Methods
-    # =========================================================================
-
-    def _get_source_content(self, project_id: str, source_id: str) -> str:
-        """
-        Get source content for marketing strategy generation.
-
-        Educational Note: Same pattern as other studio services - sample chunks
-        for large sources, use full content for small sources.
-        """
-        try:
-            from app.services.source_services import source_service
-
-            source = source_service.get_source(project_id, source_id)
-            if not source:
-                return "Error: Source not found"
-
-            # Get processed content
-            sources_dir = get_sources_dir(project_id)
-            processed_path = os.path.join(sources_dir, "processed", f"{source_id}.txt")
-
-            if not os.path.exists(processed_path):
-                return f"Source: {source.get('name', 'Unknown')}\n(Content not yet processed)"
-
-            with open(processed_path, "r", encoding="utf-8") as f:
-                full_content = f.read()
-
-            # If content is small enough, use it all
-            if len(full_content) < 15000:  # ~3750 tokens
-                return full_content
-
-            # For large sources, sample chunks
-            chunks_dir = os.path.join(sources_dir, "chunks", source_id)
-            if not os.path.exists(chunks_dir):
-                return full_content[:15000] + "\n\n[Content truncated...]"
-
-            # Get all chunks
-            chunk_files = sorted([
-                f for f in os.listdir(chunks_dir)
-                if f.endswith(".txt") and f.startswith(source_id)
-            ])
-
-            if not chunk_files:
-                return full_content[:15000] + "\n\n[Content truncated...]"
-
-            # Sample up to 10 chunks evenly distributed
-            max_chunks = 10
-            if len(chunk_files) <= max_chunks:
-                selected_chunks = chunk_files
-            else:
-                step = len(chunk_files) / max_chunks
-                selected_chunks = [chunk_files[int(i * step)] for i in range(max_chunks)]
-
-            # Read selected chunks
-            sampled_content = []
-            for chunk_file in selected_chunks:
-                chunk_path = os.path.join(chunks_dir, chunk_file)
-                with open(chunk_path, "r", encoding="utf-8") as f:
-                    sampled_content.append(f.read())
-
-            return "\n\n".join(sampled_content)
-
-        except Exception as e:
-            print(f"[MarketingStrategyAgent] Error getting source content: {e}")
-            return f"Error loading source content: {str(e)}"
-
     def _save_execution(
         self,
         project_id: str,
@@ -471,7 +185,7 @@ Please create a complete marketing strategy following the workflow:
         started_at: str,
         source_id: str
     ) -> None:
-        """Save execution log using message_service."""
+        """Save execution log for debugging."""
         message_service.save_agent_execution(
             project_id=project_id,
             agent_name=self.AGENT_NAME,
